[artf1083711]: Enable P2P feature for PN7160 chipset.
For PN220 Common MW chipset, P2P is disabled.
it is not supported for PN7220 chipset.
For PN7160 enable P2P function is used.
SystemBarUtils is removed and reveted to old aosp commit
As, there is bug in aosp and lead to nfc crash when p2p enable.
This commit also include fix for image & video transfer.

Change-Id: I8275811eed5096eb3c09cefa89f00f5ed444547b